fix(stock_zh_a_tick_tx.py): fix stock_zh_a_tick_tx_js interface
fix stock_zh_a_tick_tx_js interface
