+++ docs/data/stock/stock.md
@@ -1290,72 +1290,6 @@ date
 1900-01-01    1.0000000000000000
 ```
 
-##### 历史行情数据-网易
-
-P.S. 该接口即将移除
-
-接口: stock_zh_a_hist_163
-
-目标地址: http://quotes.money.163.com/trade/lsjysj_601318.html?year=2022&season=2
-
-描述: 网易财经-行情首页-沪深 A 股-每日行情; 该接口主要用户获取流通市值、总市值等指标
-
-限量: 单次返回指定沪深 A 股（不包含北交所）上市公司指定日期间的历史行情日频率数据, 该接口只返回未复权数据
-
-输入参数
-
-| 名称         | 类型  | 描述                                                                        |
-|------------|-----|---------------------------------------------------------------------------|
-| symbol     | str | symbol='sh601318'; 股票代码可以在 **ak.stock_zh_a_spot()** 中获取; 其中 sh689009 没有数据 |
-| start_date | str | start_date='20201103'; 开始查询的日期                                            |
-| end_date   | str | end_date='20201116'; 结束查询的日期                                              |
-
-输出参数
-
-| 名称   | 类型      | 描述      |
-|------|---------|---------|
-| 日期   | object  | -       |
-| 股票代码 | object  | -       |
-| 名称   | object  | -       |
-| 收盘价  | float64 | -       |
-| 最高价  | float64 | -       |
-| 最低价  | float64 | -       |
-| 开盘价  | float64 | -       |
-| 前收盘  | float64 | -       |
-| 涨跌额  | float64 | -       |
-| 涨跌幅  | float64 | 注意单位: % |
-| 换手率  | float64 | 注意单位: % |
-| 成交量  | float64 | 注意单位: 股 |
-| 成交金额 | float64 | 注意单位: 元 |
-| 总市值  | float64 | 注意单位: 元 |
-| 流通市值 | float64 | 注意单位: 元 |
-
-接口示例
-
-```python
-import akshare as ak
-
-stock_zh_a_hist_163_df = ak.stock_zh_a_hist_163(symbol="sh601318", start_date="20210101", end_date="20220101")
-print(stock_zh_a_hist_163_df)
-```
-
-数据示例
-
-```
-        日期    股票代码    名称  ...          成交金额        总市值        流通市值
-0    2021-01-04  601318  中国平安  ...  8.814287e+09  1.557111e+12  9.227264e+11
-1    2021-01-05  601318  中国平安  ...  8.795810e+09  1.540476e+12  9.128686e+11
-2    2021-01-06  601318  中国平安  ...  7.950362e+09  1.567348e+12  9.287927e+11
-3    2021-01-07  601318  中国平安  ...  9.137581e+09  1.576854e+12  9.344256e+11
-4    2021-01-08  601318  中国平安  ...  1.006906e+10  1.573929e+12  9.326924e+11
-..          ...     ...   ...  ...           ...           ...           ...
-238  2021-12-27  601318  中国平安  ...  1.373482e+09  9.136465e+11  5.414166e+11
-239  2021-12-28  601318  中国平安  ...  3.137431e+09  9.299159e+11  5.510576e+11
-240  2021-12-29  601318  中国平安  ...  1.913670e+09  9.288191e+11  5.504077e+11
-241  2021-12-30  601318  中国平安  ...  2.167180e+09  9.215070e+11  5.460746e+11
-242  2021-12-31  601318  中国平安  ...  1.805071e+09  9.215070e+11  5.460746e+11
-```
-
 ##### 分时数据-新浪
 
 接口: stock_zh_a_minute
@@ -1631,64 +1565,6 @@ print(stock_zh_a_tick_tx_js_df)
 1407  14:59:50  3.52  0.01      69   24288   买盘
 1408  14:59:59  3.51 -0.01     100   35100   卖盘
 1409  15:00:02  3.52  0.01      10    3520   买盘
-```
-
-##### 网易财经
-
-P.S. 该接口暂时不能使用
-
-接口: stock_zh_a_tick_163
-
-目标地址: http://quotes.money.163.com/trade/cjmx_000001.html#01b05(示例)
-
-描述: A 股 Tick 数据是从网易财经获取, 历史数据按日频率更新, 晚上 10 更新数据; 可以调用 **ak.stock_zh_a_tick_163_now(
-symbol="000001")** 接口获取
-当日数据(该接口只能在交易日获取数据), 主要参数无市场标识
-
-限量: 单次返回具体某个 A 上市公司的近 5 个交易日的历史分笔行情数据; 该接口目前还不支持北交所的股票; 该接口目前不能返回数据
-
-输入参数
-
-| 名称         | 类型       | 描述                    |
-|------------|----------|-----------------------|
-| symbol     | str      | symbol="sh600000";    |
-| trade_date | datetime | trade_date="20200408" |
-
-输出参数
-
-| 名称   | 类型      | 描述      |
-|------|---------|---------|
-| 时间   | object  | -       |
-| 成交价  | float64 | -       |
-| 价格变动 | float64 | -       |
-| 成交量  | int64   | 注意单位: 手 |
-| 成交额  | float64 | 注意单位: 元 |
-| 性质   | object  | 买卖盘标记   |
-
-接口示例
-
-```python
-import akshare as ak
-
-stock_zh_a_tick_163_df = ak.stock_zh_a_tick_163(symbol="sh600848", trade_date="20210127")
-print(stock_zh_a_tick_163_df)
-```
-
-数据示例
-
-```
-      成交时间    成交价  价格变动  成交量     成交额   性质
-0     09:25:03  19.48  0.13   39   75972   买盘
-1     09:30:02  19.47 -0.01   17   33095  中性盘
-2     09:30:06  19.37 -0.10    6   11667   卖盘
-3     09:30:12  19.40 -0.06    7   13580  中性盘
-4     09:30:15  19.40  0.00    2    3880   买盘
-        ...    ...   ...  ...     ...  ...
-1306  14:56:40  19.32 -0.01   15   28980   卖盘
-1307  14:56:46  19.32  0.00   69  133308   卖盘
-1308  14:56:52  19.31 -0.01    7   13523   卖盘
-1309  14:57:02  19.32  0.01    6   11592   买盘
-1310  15:00:03  19.32  0.00  299  577668   买盘
 ```
 
 ### A股-CDR
